+++ templates/starter-template/js/init.js
@@ -1,7 +1,7 @@
 (function($){
   $(function(){
 
-    $('.button-collapse').sideNav();
+    $('.button-collapse').sidenav();
 
   }); // end of document ready
 })(jQuery); // end of jQuery name space